# openapi.yaml
openapi: 3.0.3
info:
  title: Dog Trainers Directory API
  description: |
    API for the Dog Trainers Directory platform.
    
    This API provides endpoints for:
    - Public access to trainer listings and search
    - Trainer profile management
    - Admin moderation and management
    - Emergency triage and contacts
    - Authentication with MFA support
    
    ## Authentication
    
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    
    `Authorization: Bearer <access_token>`
    
    ## Rate Limiting
    
    All endpoints are rate limited. Rate limit headers are included in responses:
    
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    - `Retry-After`: Seconds until retry (only when rate limited)
  version: 1.0.0
  contact:
    name: Dog Trainers Directory Support
    email: support@dogtrainersdirectory.com.au
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.dogtrainersdirectory.com.au/v1
    description: Production server
  - url: https://staging-api.dogtrainersdirectory.com.au/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
...
paths:
...
  /trainer/issues:
    post:
      tags:
        - Trainers
      summary: Add issue
      description: Add a behavior issue to trainer profile
      operationId: addIssue
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddIssueRequest'
      responses:
        '200':
          description: Issue added successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddIssueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /trainer/issues/{behaviorIssue}:
    delete:
      tags:
        - Trainers
      summary: Remove issue
      description: Remove a behavior issue from trainer profile
      operationId: removeIssue
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/BehaviorIssue'
      responses:
        '200':
          description: Issue removed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemoveIssueResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /trainer/featured/purchase:
    post:
      tags:
        - Trainers
      summary: Purchase featured placement
      description: Purchase featured placement for trainer profile
      operationId: purchaseFeatured
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurchaseFeaturedRequest'
      responses:
        '200':
          description: Featured placement purchased successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseFeaturedResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /trainer/featured/status:
    get:
      tags:
        - Trainers
      summary: Check featured status
      description: Check featured placement status for trainer
      operationId: getFeaturedStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CouncilId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturedStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # ============================================================================
  # ADMIN ENDPOINTS
  # ============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  parameters:
...
    CouncilId:
      name: council_id
      in: query
      description: Filter by council ID
      required: false
      schema:
        type: string
        format: uuid
...
    BehaviorIssue:
      name: behavior_issue
      in: query
      description: Filter by behavior issue
      required: false
      schema:
        type: string
        enum:
          - Pulling on lead
          - Separation anxiety
          - Barking
          - Aggression
          - Jumping up on people
          - Recall issues
          - Socialisation
          - Chewing
          - Digging
          - House training
          - Fear/phobias
          - Other
...
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
        version:
          type: string
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
...
    Trainer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        resource_type:
          type: string
          enum: [trainer, behaviour_consultant, emergency_vet, urgent_care, emergency_shelter]
        locality_id:
          type: string
          format: uuid
        council_id:
          type: string
          format: uuid
        phone:
          type: string
        email:
          type: string
          nullable: true
        website:
          type: string
          format: uri
          nullable: true
        description:
          type: string
          nullable: true
        age_specialties:
          type: array
          items:
            type: string
        behavior_issues:
          type: array
          items:
            type: string
        service_type_primary:
          type: string
          nullable: true
        service_type_secondary:
          type: array
          items:
            type: string
          nullable: true
        abr_abn:
          type: string
          nullable: true
        emergency_phone:
          type: string
          nullable: true
        emergency_hours:
          type: string
          nullable: true
        emergency_services:
          type: array
          items:
            type: string
          nullable: true
        verified:
          type: boolean
        claimed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FeaturedPlacement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        business_id:
          type: string
          format: uuid
        council_id:
          type: string
          format: uuid
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        status:
          type: string
          enum: [queued, active, expired, refunded, cancelled]
        queue_position:
          type: integer
          nullable: true
        refund_reason:
          type: string
          nullable: true
        refunded_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AddIssueRequest:
      type: object
      required:
        - behavior_issue
      properties:
        behavior_issue:
          type: string
          enum:
            - Pulling on lead
            - Separation anxiety
            - Barking
            - Aggression
            - Jumping up on people
            - Recall issues
            - Socialisation
            - Chewing
            - Digging
            - House training
            - Fear/phobias
            - Other
    AddIssueResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trainer:
              $ref: '#/components/schemas/Trainer'
            added_issue:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    RemoveIssueResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            trainer:
              $ref: '#/components/schemas/Trainer'
            removed_issue:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    PurchaseFeaturedRequest:
      type: object
      required:
        - council_id
        - duration_days
        - payment_method_id
      properties:
        council_id:
          type: string
          format: uuid
        duration_days:
          type: integer
          minimum: 1
          maximum: 365
        payment_method_id:
          type: string
          minLength: 1
    PurchaseFeaturedResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            featured_placement:
              $ref: '#/components/schemas/FeaturedPlacement'
            payment_intent_id:
              type: string
            client_secret:
              type: string
            amount:
              type: integer
            currency:
              type: string
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    FeaturedStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            featured:
              $ref: '#/components/schemas/FeaturedPlacement'
            queue_position:
              type: integer
              nullable: true
            can_purchase:
              type: boolean
            next_available_date:
              type: string
              format: date-time
              nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'